nagios_check_certificate/requirements.txt
certifi==2020.4.5.2
pycurl==7.43.0.5
pyOpenSSL==19.1.0